backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary # For PostgreSQL
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
